build 200604. sound loading logging enhancement, driving aid throttle data enhancement, external camera movement control fix

+++ Logs.cpp
@@ -51,7 +51,7 @@ std::string filename_scenery() {
 void WriteLog( const char *str, logtype const Type ) {
 
     if( str == nullptr ) { return; }
-    if( true == TestFlag( Global.DisabledLogTypes, Type ) ) { return; }
+    if( true == TestFlag( Global.DisabledLogTypes, static_cast<unsigned int>( Type ) ) ) { return; }
 
     if (Global.iWriteLogEnabled & 1) {
         if( !output.is_open() ) {
@@ -79,7 +79,7 @@ void WriteLog( const char *str, logtype const Type ) {
 void ErrorLog( const char *str, logtype const Type ) {
 
     if( str == nullptr ) { return; }
-    if( true == TestFlag( Global.DisabledLogTypes, Type ) ) { return; }
+    if( true == TestFlag( Global.DisabledLogTypes, static_cast<unsigned int>( Type ) ) ) { return; }
 
     if (!errors.is_open()) {
